+++ src/components/create-post/create-post.css
@@ -40,16 +40,3 @@
     cursor: pointer;
     color: var(--primary);
 }
-.post-btn{
-    width: max-content;
-    padding: 0.5rem 1.5rem;
-    background: var(--primary);
-    color: white;
-    outline: none;
-    border: 1px solid var(--primary);
-    border-radius: 1px;
-    cursor: pointer;
-}
-.post-btn:hover{
-    background: var(--secondary);
-}
