+++ src/pages/homepage/homepage.css
@@ -3,7 +3,6 @@
     background: var(--background-grey);
     text-align: left;
     padding: 2rem 5rem;
-    /* height: 100vh; */
     justify-content: space-between;
     box-sizing: border-box;
 }
