+++ src/styles.css
@@ -1,5 +1,6 @@
 @import url("./components/user-avatar/user-avatar.css");
 @import url("./pages/homepage/homepage.css");
+@import url("./components/create-post/create-post.css");
 
 :root{
     --background-grey: #F7F7F7;
@@ -11,6 +12,19 @@
     padding: 1rem 2rem;
     border-radius: 4px;
     width: 100%;
+}
+.post-btn{
+    width: max-content;
+    padding: 0.5rem 1.5rem;
+    background: var(--primary);
+    color: white;
+    outline: none;
+    border: 1px solid var(--primary);
+    border-radius: 1px;
+    cursor: pointer;
+}
+.post-btn:hover{
+    background: var(--secondary);
 }
 .avatar{
     width: 2.5rem;
